Added hover to reveal interactive object
Interactive objects now change layer when hovered over, causing them to be rendered in an overlay camera instead of the main camera.

There is a new overlay camera and layer, plus several changes to the renderer, player prefab and interactable prefabs to facilitate this update. There may need to be additional changes so that the overlay camera's rendering better matches the main camera.