Cursor overhaul
Instead of using an image to display the cursor, a new script has been added that makes use of Unity's own custom cursor system.
This overhaul allows the cursor's size and offset to be consistent with that of the user's system.

Related cursor sprites and prefabs have been modified for this overhaul.