Attached camera switch UI to CameraCinemaSwitch.cs
Camera switch UI visibility is now directly controlled by CameraCinemaSwitch.cs, so that there are no more issues with inconsistent UI visibility.

This also has allowed the number key hotfix to be reverted. It is now possible to switch from special camera to a standard camera, without breaking the UI.