Added back in end day button to HideUI script
endDayButtonFade variable and script functionality have been added back into HideUIOnMouseOver.cs. This part of the script was necessary as if the HUD was hidden you wouild just end up with a really jarring EndDay button floating on screen. I have increased the collider arround the button to make it easier to click without triggering the fade effect.