Added scroll to move object
Scroll wheel can now move an object forward and backward. This now uses multiple offset variables to track where an object should be, relative to the camera and where the cursor pointing towards in the scene. Still has a few kinks to sort out, namely that scrolling the literal offset can go past the visual limit.