Added BasketballHoop.cs
BasketballHoop.cs allows an Interactable object to move to a position after hitting a trigger. This script has been added to the bins, making it possible to physically toss trash into the bin instead of clicking.

Also modified PRE_BIN.prefab for this purpose. The bin's PlacementSpot is now in the prefab, as BasketballHoop.cs requires a PlacementSpot.cs as a reference.