Fixed jittery object movement
Fixed an oversight with the interactable object movement, which results in jittery collisions. Object now uses MovePosition() in tandem with MoveTowards() for this smoother movement.

Also added delta time to movement, so that it is consistent across different frame rates. Movement when an object is near its final position is also smoother now, for further reduced jitter.