Cursor state now self-controlled
Cursor state is now directly controlled by the cursor script for consistency, and to remove faulty dependency on OnHoverEnter/OnHoverExit.

Cursor state is now controlled by a raycast that checks if the mouse is pointing at an object in an interaction layer. There is also special behavior for layers to be ignored when an object is or is not selected, as well as special BasketballHoop behavior.

Previous code lines in other scripts that affect the cursor state have been commented out.

Also moved InteractableStateController.cs to a different folder.